Use jinja for payload generation
